Create reports subfolder under domain
