+++ src/compositionRoot.ts
@@ -4,24 +4,24 @@ import { NHWADataApprovalDefaultRepository } from "./data/NHWADataApprovalDefaul
 import { MALDataDuplicationDefaultRepository } from "./data/MALDataDuplicationDefaultRepository";
 import { NHWADataCommentsDefaultRepository } from "./data/NHWADataCommentsDefaultRepository";
 import { WIDPAdminDefaultRepository } from "./data/WIDPAdminDefaultRepository";
-import { GetWIDPAdminDefaultUseCase } from "./domain/admin/usecases/GetWIDPAdminDefaultUseCase";
-import { SaveWIDPAdminDefaultCsvUseCase } from "./domain/admin/usecases/SaveWIDPAdminDefaultCsvUseCase";
+import { GetWIDPAdminDefaultUseCase } from "./domain/reports/admin/usecases/GetWIDPAdminDefaultUseCase";
+import { SaveWIDPAdminDefaultCsvUseCase } from "./domain/reports/admin/usecases/SaveWIDPAdminDefaultCsvUseCase";
 import { GetConfig } from "./domain/common/usecases/GetConfig";
 import { GetOrgUnitsUseCase } from "./domain/common/usecases/GetOrgUnitsUseCase";
-import { UpdateStatusUseCase } from "./domain/nhwa-approval-status/usecases/CompleteDataSetsUseCase";
-import { GetApprovalColumnsUseCase } from "./domain/nhwa-approval-status/usecases/GetApprovalColumnsUseCase";
-import { GetDataSetsUseCase } from "./domain/nhwa-approval-status/usecases/GetDataSetsUseCase";
-import { SaveApprovalColumnsUseCase } from "./domain/nhwa-approval-status/usecases/SaveApprovalColumnsUseCase";
-import { SaveDataSetsUseCase } from "./domain/nhwa-approval-status/usecases/SaveDataSetsCsvUseCase";
-import { GetDataValuesUseCase } from "./domain/nhwa-comments/usecases/GetDataValuesUseCase";
-import { SaveDataValuesUseCase } from "./domain/nhwa-comments/usecases/SaveDataValuesCsvUseCase";
-import { UpdateStatusAndDuplicateUseCase } from "./domain/mal-dataset-duplication/usecases/CompleteAndDuplicateDataSetsUseCase";
-import { GetApprovalAndDuplicateColumnsUseCase } from "./domain/mal-dataset-duplication/usecases/GetApprovalAndDuplicateColumnsUseCase";
-import { GetDataSetsDuplicationUseCase } from "./domain/mal-dataset-duplication/usecases/GetDataSetsDuplicationUseCaseOptions";
-import { SaveApprovalAndDuplicateColumnsUseCase } from "./domain/mal-dataset-duplication/usecases/SaveApprovalDuplicateColumnsUseCase";
-import { SaveDataSetsDuplicationUseCase } from "./domain/mal-dataset-duplication/usecases/SaveDataSetsDuplicationUseCase";
+import { UpdateStatusUseCase } from "./domain/reports/nhwa-approval-status/usecases/CompleteDataSetsUseCase";
+import { GetApprovalColumnsUseCase } from "./domain/reports/nhwa-approval-status/usecases/GetApprovalColumnsUseCase";
+import { GetDataSetsUseCase } from "./domain/reports/nhwa-approval-status/usecases/GetDataSetsUseCase";
+import { SaveApprovalColumnsUseCase } from "./domain/reports/nhwa-approval-status/usecases/SaveApprovalColumnsUseCase";
+import { SaveDataSetsUseCase } from "./domain/reports/nhwa-approval-status/usecases/SaveDataSetsCsvUseCase";
+import { GetDataValuesUseCase } from "./domain/reports/nhwa-comments/usecases/GetDataValuesUseCase";
+import { SaveDataValuesUseCase } from "./domain/reports/nhwa-comments/usecases/SaveDataValuesCsvUseCase";
+import { UpdateStatusAndDuplicateUseCase } from "./domain/reports/mal-dataset-duplication/usecases/CompleteAndDuplicateDataSetsUseCase";
+import { GetApprovalAndDuplicateColumnsUseCase } from "./domain/reports/mal-dataset-duplication/usecases/GetApprovalAndDuplicateColumnsUseCase";
+import { GetDataSetsDuplicationUseCase } from "./domain/reports/mal-dataset-duplication/usecases/GetDataSetsDuplicationUseCaseOptions";
+import { SaveApprovalAndDuplicateColumnsUseCase } from "./domain/reports/mal-dataset-duplication/usecases/SaveApprovalDuplicateColumnsUseCase";
+import { SaveDataSetsDuplicationUseCase } from "./domain/reports/mal-dataset-duplication/usecases/SaveDataSetsDuplicationUseCase";
 import { D2Api } from "./types/d2-api";
-import { GetDataDiffUseCase } from "./domain/mal-dataset-duplication/usecases/GetDataDiffUseCase";
+import { GetDataDiffUseCase } from "./domain/reports/mal-dataset-duplication/usecases/GetDataDiffUseCase";
 import { getReportType } from "./webapp/utils/reportType";
 
 export function getCompositionRoot(api: D2Api) {
