--- src/domain/admin/usecases/SaveWIDPAdminDefaultCsvUseCase.ts
@@ -1,10 +1,0 @@
-import { MetadataObject } from "../../common/entities/MetadataObject";
-import { WIDPAdminRepository } from "../repositories/WIDPAdminRepository";
-
-export class SaveWIDPAdminDefaultCsvUseCase {
-    constructor(private metadataRepository: WIDPAdminRepository) {}
-
-    async execute(filename: string, metadata: MetadataObject[]): Promise<void> {
-        this.metadataRepository.save(filename, metadata);
-    }
-}
