--- src/domain/mal-dataset-duplication/usecases/SaveApprovalDuplicateColumnsUseCase.ts
@@ -1,10 +1,0 @@
-import { UseCase } from "../../../compositionRoot";
-import { MALDataDuplicationRepository } from "../repositories/MALDataDuplicationRepository";
-
-export class SaveApprovalAndDuplicateColumnsUseCase implements UseCase {
-    constructor(private approvalRepository: MALDataDuplicationRepository) {}
-
-    execute(namespace: string, columns: string[]): Promise<void> {
-        return this.approvalRepository.saveColumns(namespace, columns);
-    }
-}
