--- src/domain/nhwa-approval-status/usecases/SaveApprovalColumnsUseCase.ts
@@ -1,10 +1,0 @@
-import { UseCase } from "../../../compositionRoot";
-import { NHWADataApprovalRepository } from "../repositories/NHWADataApprovalRepository";
-
-export class SaveApprovalColumnsUseCase implements UseCase {
-    constructor(private approvalRepository: NHWADataApprovalRepository) {}
-
-    execute(columns: string[]): Promise<void> {
-        return this.approvalRepository.saveColumns(columns);
-    }
-}
