+++ src/domain/reports/mal-dataset-duplication/usecases/GetApprovalAndDuplicateColumnsUseCase.ts
@@ -1,0 +1,10 @@
+import { UseCase } from "../../../../compositionRoot";
+import { MALDataDuplicationRepository } from "../repositories/MALDataDuplicationRepository";
+
+export class GetApprovalAndDuplicateColumnsUseCase implements UseCase {
+    constructor(private approvalRepository: MALDataDuplicationRepository) {}
+
+    execute(namespace: string): Promise<string[]> {
+        return this.approvalRepository.getColumns(namespace);
+    }
+}
