+++ src/domain/reports/nhwa-approval-status/entities/DataApprovalItem.ts
